% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_regres.R
\name{estim_regres}
\alias{estim_regres}
\title{Determine point estimate, SD and SE, 95\% Credibility Intervals, and interval width, for both
differences in raw means and Cohen's D's for multiple sample sizes}
\usage{
estim_regres(data, vars_of_interest, k, row_select, name)
}
\arguments{
\item{data}{Dataframe with the data to be analyzed}

\item{vars_of_interest}{Vector containing the names of the variables to be compared on their means}

\item{k}{The number of permutations to be used for each sample size}

\item{row_select}{The range of sample size to be used}

\item{name}{The title of the dataset or variables to be displayed with the figure}
}
\value{
And array containing the difference in raw means with associated SD, SE, 95\% CI, and width of
the 95\% CI, the Cohen's D value of the difference with associated SD, SE, 95\% CI, and width of
the 95\% CI, ans the sample size on which each row of results is based.
}
\description{
Determine point estimate, SD and SE, 95\% Credibility Intervals, and interval width, for both
differences in raw means and Cohen's D's for multiple sample sizes
}
