% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_corr.R
\name{sample_corr}
\alias{sample_corr}
\title{helper function to sample the data and estimate the correlation}
\usage{
sample_corr(data, vars_of_interest, sample_size)
}
\arguments{
\item{data}{Dataframe with the data to be analyzed}

\item{vars_of_interest}{Vector containing the names of the variables to be compared on their means}

\item{sample_size}{The range of sample size to be used}
}
\value{
list with correlations and credible intervals for given sample_size
}
\description{
helper function to sample the data and estimate the correlation
}
