% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{self_eval}
\alias{self_eval}
\title{Self-evaluations task fMRI region of interest data}
\format{
A data frame with 149 rows and 4 variables:
\describe{
\item{participant_id}{unique id for every participant}
\item{age}{age in years (11.00-20.92)}
\item{mpfc_self}{parameter estimates for the left medial prefrontal cortex during
self-evaluation (-2.82-4.97)}
\item{mpfc_control}{parameter estimates for the lmedial prefrontal cortex during
the control condition (-7.17-3.50)}
}
}
\source{
{van der Cruijsen, R., Blankenstein, N. E., Spaans, J. P., Peters, S., & Crone, E. A.
(2023). Longitudinal self-concept development in adolescence. Social Cognitive and Affective
Neuroscience, 18(1), nsac062.}
\doi{10.1093/scan/nsac062}
}
\usage{
self_eval
}
\description{
A dataset containing the parameter estimates of the
left medial prefrontal cortex (x = −6, y = 50, z = 4),
with one value for the mean activation during self-evaluation
and one value for the mean activation during the control condition for all participants.
}
\keyword{datasets}
