% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{theta}
\alias{theta}
\title{Add together two numbers}
\usage{
theta(x)
}
\arguments{
\item{x}{A data vector}
}
\value{
The ratio of \code{mean(x)} and \code{sd(x)}
}
\description{
Add together two numbers
}
\examples{
theta(rnorm(50,0,1))
}
