% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vicar_char}
\alias{vicar_char}
\title{Vicarious Charity task fMRI region of interest data}
\format{
A data frame with 156 rows and 4 variables:
\describe{
\item{participant_id}{unique id for every participant}
\item{age}{age in years (11.00-21.17)}
\item{nacc_selfgain}{parameter estimates for the left nucleus accumbens during
gaining for self (-5.66-3.05)}
\item{nacc_bothnogain}{parameter estimates for the left nucleus accumbens during
no-gain for self and charity (-6.44-2.97)}
}
}
\source{
{Spaans, J., Peters, S., Becht, A., van der Cruijsen, R., van de Groep, S.,
& Crone, E. A. (2023). Longitudinal neural and behavioral trajectories of charity
contributions across adolescence. Journal of Research on Adolescence, 33(2), 480–495.}
\doi{10.1111/jora.12820}
}
\usage{
vicar_char
}
\description{
A dataset containing the parameter estimates from
the anatomical mask of the left nucleus accumbens (Harvard-Oxford subcortical atlas;
thresholded at 40\%; center-of-mass coordinates x = −10, y = 12, z = −7; 28 voxels included),
with one value for the mean activation during gaining for self and one value for the mean
activation during no-gain for self and charity for all participants.
}
\keyword{datasets}
