% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_corr.R
\name{estim_corr}
\alias{estim_corr}
\title{Estimate correlations}
\usage{
estim_corr(data, vars_of_interest, k, sample_size, name)
}
\arguments{
\item{data}{Dataframe with the data to be analyzed}

\item{vars_of_interest}{Vector containing the names of the variables to be
correlated}

\item{k}{The number of permutations to be used for each sample size}

\item{sample_size}{The range of sample size to be used (min:max)}

\item{name}{The title of the dataset or variables to be displayed with the
figure}
}
\value{
\itemize{
\item \code{tbl_selection} returns a \code{\link[tibble:tibble]{tibble::tibble()}} containing estimates of the Pearson’s correlation between
two correlated variables for five different sample sizes (starting with the
minimum sample size, then 1/5th parts of the total dataset).
\item \code{fig_corr} returns a scatterplot where for the five different sample sizes, 10 out of the total number
of HDCIs computed are displayed (in green). The average estimate with
credible interval summarizing the total number of HDCIs for each sample size
are plotted in orange
\item \code{fig_corr_nozero} returns a barplot where for each of the five sample sizes the proportion of permutations
not containing zero is displayed
\item \code{tbl_total} returns a \code{\link[tibble:tibble]{tibble::tibble()}} containing estimates of the Pearson’s correlation between
two correlated variables for all sample sizes, including the lower and upper
range and the permuation number.
}
}
\description{
\code{estim_corr} determines point estimate, SD and SE, 95\\% Credibility
Intervals, and interval width, for Pearson correlations for multiple sample
sizes
}
\examples{
data_gambling <- gambling
estim_corr(data_gambling, 
  c("lnacc_self_winvsloss", "age"), 10, 
  20:221, "Gambling NAcc correlation with age")
}
